Moved class App to App.js
